Create first section of the README
This app will be created using a process called Readme Driven Development. Read more about RDD here: http://tom.preston-werner.com/2010/08/23/readme-driven-development.html. 

This is the first step of this application build. With this section of the README completed, the sections outlined in this commit will be the first part to be worked on in the codebase.

This commit creates the instructions for how to install the application.